requirements.txt
absl-py==2.1.0
actionlib==1.14.0
aiofiles==0.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alabaster==0.7.16
angles==1.9.13
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
apiage==0.1.4
asttokens==3.0.0
async-timeout==4.0.3
asyncio==3.4.3
atomicwrites==1.4.1
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bokeh==3.1.1
boltons==25.0.0
bondpy==1.8.6
bs4==0.0.1
build==1.2.2.post1
cachetools==5.5.0
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
casadi==3.6.7
catkin==0.8.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==2.1.1
click==7.1.2
cloudpickle==3.1.0
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cryptography==45.0.5
cv-bridge==1.16.2
cycler==0.12.1
dataclasses-json==0.6.7
decorator==4.4.2
Deprecated==1.2.5
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.9.0
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.21.2
dynamic-reconfigure==1.7.3
easydict==1.13
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.1.0
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
furl==2.1.4
fusepy==3.0.1
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
git-filter-repo==2.47.0
gitdb==4.0.12
gitdb2==4.0.2
GitPython==2.1.11
gltf==0.7.22
google-auth==2.37.0
google-auth-oauthlib==1.0.0
goto-statement==1.2
gpgrecord==0.0.4
greenlet==3.1.1
grpcio==1.69.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
h11==0.16.0
h5py==3.11.0
hanziconv==0.3.2
httpcore==1.0.9
httpx==0.28.1
idna==3.10
image-geometry==1.16.2
imageio==2.35.1
imageio-ffmpeg==0.5.1
imagesize==1.4.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
interactive-markers==1.12.0
ipython==7.3.0
jedi==0.19.2
Jinja2==2.11.3
jiter==0.9.1
joblib==1.4.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langsmith==0.1.147
laser_geometry==1.6.7
llvmlite==0.41.1
lxml==6.0.0
Markdown==3.7
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.7.5
matplotlib-inline==0.1.7
mediapy==1.2.4
message-filters==1.17.0
metadir==0.0.2
metadrive-simulator==0.4.3
metaform==1.0.2.4
metatype==0.1.3.7
metawiki==0.1.13
mistune==3.1.3
more-itertools==10.7.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
mypy_extensions==1.1.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.82.1
opencv-python==4.10.0.84
openpyxl==3.1.5
orderedmultidict==1.0.1
orjson==3.10.15
packaging==24.2
Panda3D==1.10.13
panda3d-gltf==0.13
panda3d-simplepbr==0.12.0
pandas==1.3.5
paramiko==2.10.1
parse==1.20.2
parso==0.8.4
pettingzoo==1.24.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip-tools==7.4.1
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.6.0
proglog==0.1.10
progress==1.6.1
progressbar==2.5
prompt-toolkit==2.0.10
propcache==0.2.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.42
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.1
pymongo==3.7.2
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==3.1.4
pyproj==3.5.0
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.15.0
pyquaternion==0.9.9
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.6.8
pytest==4.4.1
python-dateutil==2.8.0
python-gnupg==0.5.4
python-qt-binding==0.4.4
python-slugify==8.0.4
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
rdkit==2024.3.5
requests==2.28.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
resource_retriever==1.12.7
rosbag==1.17.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.17.0
roslaunch==1.17.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.17.0
rosmake==1.15.8
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospy==1.17.0
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.8
roswtf==1.17.0
rqt-console==0.4.12
rqt-image-view==0.4.17
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rsa==4.9
rviz==1.14.25
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
selenium==3.141.0
selenium-wire==2.1.1
sensor-msgs==1.13.1
sentry-sdk==2.20.0
setproctitle==1.3.4
shapely==2.0.6
six==1.17.0
slumber==0.7.1
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
soundfile==0.13.0
soupsieve==2.7
Sphinx==2.0.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
stable_baselines3==2.4.1
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
text-unidecode==1.3
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
threadpoolctl==3.5.0
tianshou==0.5.1
tinydb==4.8.2
tomli==2.2.1
topic-tools==1.17.0
torch==2.4.1
tornado==6.4.2
tqdm==4.31.1
traitlets==5.14.3
triton==3.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
typology==0.2.3
tzdata==2024.2
urllib3==1.26.5
wandb==0.19.3
wcwidth==0.2.13
Werkzeug==3.0.6
wrapt==1.17.2
xacro==1.14.18
xarray==0.12.1
xyzservices==2024.9.0
yapf==0.43.0
yarl==1.15.2
yolk3k==0.9
zipp==3.20.2
